+++ src/utils/constants.ts
@@ -10,3 +10,10 @@ interface Menu {
 export const menus: Menu[] = [
   { id: 'home', name: 'Home', icon: Grid3x3, isActive: true },
 ];
+
+
+export const CLI_VERSION = '0.0.13';
+
+export const ONE_DAY_MS = 24 * 60 * 60 * 1000;
+
+export const ONE_HOUR_MS = 2 * 60 * 1000; // update to 2 min for fast update